test(iosApp.StopDetailsFilteredView): enhanced favorites flow
